Changed `with(…)` to always return same type, and new `withMap(…)` to handle arbitrary return
I was experiencing issues in my own code with Swift's type inference getting confused as to which `with(…)` variant to call, and 99% of the time the solution being to typecast the result of `with(…) { … }` to the desired (the subject's) type.  Many of these were because we were performing `with(…)` on `AType` but then returning it into a variable of type `AType?` or `AType!`.

I experimented with creating `Optional` specializations of `with(…)` and renaming the value-type versions of `with(…)`, but eventually discovered that because of the arbitrary `ReturnT:Any` versions of `with(…)`, the compiler wasn't sure to implicitly optional-wrap the `AType` result or to use the arbitrary `ReturnT:Any` variants of `with(…)`.

The solution was to rename the arbitrary-return variants to `withMap(…)`, which is closer to what we're doing— _“with an input of a given type, we're mapping to a different type of output”_.  In practical use, this version of `with…(…)` isn't used as much, so I guess having it under a different name is fine.

It's unfortunate that this seems to be necessary— I'd really like everything to be under one function name.  And it's weird too— `with(AType()) { … }` is basically just syntactic sugar for the built-in `{ … }(AType())` create-anonymouse-closure-and-call-it syntax, and the latter seems to work everywhere in all situations.  Oh well.

* Renamed the ref-type arbitrary-return form to `withMap`.
* Added a value-type arbitrary-return form, also named `withMap`.  
	▸ Originally, I didn't think this would be necessary— if you're creating a value, you'd want to hang onto it.  But after looking at usages, it became clear that the use-case of creating a value just to get info out of it, or to pass it somewhere else to be consumed is just as viable for value types as ref types.
* Updated README.