Eliminated object-type vs. value-type variants
It's become apparent to me as I had to make the object-`SubjectT` and value-`SubjectT` more and more similar in order for everything to work that I really didn't need overloaded versions of `with…(…)` for them anymore— they were both taking `inout`-arg closures, and handling the `subject` the same way.  However, I had left the distinction in since it made sense to require doing something with the return value of `with(…)` with a value-`SubjectT`, while an object-`SubjectT` didn't need that `@discardableResult`.

But there's also the use-case of creating an object/value with `with(…)` and then passing/setting it somewhere else at the end of the closure, and not needing the return value from the `with(…){}`, which is perfectly valid.

Therefore, removing the object/value distinction, leaving just the `Any` variants of `with(…)` and `withMap(…)`.

* Removed the `AnyObject` variant of `with(…)`, and added `@discardableResult` to the `Any` variant.
* Removed the `AnyObject` variant of `withMap(…)`.
* Updated the README.