Changed methods to be `@_transparent`
While `@_transparent` is an underscored (usage discouraged) attribute, it fits really well for what we're doing here.  According to the docs, _“Semantically, `@_transparent` means something like 'treat this operation as if it were a primitive operation'”_, which is exactly how we want `with(…)`/`withMap(…)` to be used in client code.

In particular, we **do want** the effect of `@_transparent` that _“Debug info SHOULD skip over the inlined operations when single-stepping through the calling function.”_— it's really annoying to step into `with(…)`/`withMap(…)`'s implementation when it doesn't do anything novel (effectively just calls the closure), and it'll be much easier to debug client code if step just moved the program counter into the client code's closure.

More details at https://github.com/apple/swift/blob/main/docs/TransparentAttr.md , but here's the “should you use `@_transparent`” checklist:

* _Does the implementation of this function ever have to change?_  
	Functionally, no.  Updates will almost certainly only be improving compiler semantics and syntax for newer Swift features.
* _Does the implementation need to call private things …?_  
	Nope.
* _Is it okay if the function is not inlined?_  
	Yes, except that we really want the “Debug info SHOULD skip over the inlined operation” mentioned above for ease of use.
* _Is it a problem if the function is inlined even under -Onone?_  
	Nope.
* _Is it a problem if you can't step through the function that's been inlined?_  
	No, that's exactly what we want.  The implementation is trivial enough that you'll never need to step into it— it really is meant to work like a language feature (found natively in other languages).
* _Is it okay if the inlining happens after all the dataflow diagnostics?_  
	Not 100% sure, but don't really care.